runtime.ConnRouter — generic conn-routing for endpoint plugins
main used to hardcode `Family == "sql"` when building the WG
forwarder's dstIP→endpoint index. Replace with a runtime interface:
endpoint plugins whose body implements ConnRouter declare the host
tuples they claim, and BuildConnIndex walks every such body
generically. Postgres implements it today; clickhouse_native or any
future binary-protocol endpoint plugs in by implementing the same
method — no main-package change required.

  // config/runtime/conn_route.go
  type ConnRouter interface {
      ConnRouteHosts() []string
  }

  func BuildConnIndex(*config.CompiledPolicy) *ConnIndex
  func (*ConnIndex) Lookup(dstIP string) []*config.CompiledEndpoint

Drops top-level postgres_index.go (postgres-specific code in main).
Gateway.pgIdx → Gateway.connIdx.
